Support Racket on Chez Scheme
Use the FFI instead of depending on Racket BC's C API.

(But we still build a shared C library, because I'm not sure
about the portability of the `struct termios` type.)

Closes https://github.com/tonyg/racket-ansi/issues/6
