+++ ansi/test-raw.rkt
@@ -1,6 +1,5 @@
 #lang racket/base
 
-(require "private/tty-raw-extension.rkt")
 (require "main.rkt")
 
 (require racket/set)
