+++ info.rkt
@@ -1,4 +1,8 @@
 #lang setup/infotab
+(define pkg-name 'ansi)
 (define collection 'multi)
-(define deps '("base" "dynext-lib" "make"))
-(define build-deps '("rackunit-lib"))
+(define deps
+  '("base"
+    "dynext-lib"))
+(define build-deps
+  '())
